[MISC] Move {core->utility}/char_operations/pretty_print.hpp
